[Tests] Cleanup.
