+++ Demo/DemoTests/ReactiveCocoaMoyaProviderTests.swift
@@ -1,7 +1,7 @@
 import Quick
 import Nimble
 import ReactiveCocoa
-import ReactiveMoya
+import Moya
 import Alamofire
 
 class ReactiveCocoaMoyaProviderSpec: QuickSpec {
@@ -212,81 +212,6 @@ class ReactiveCocoaMoyaProviderSpec: QuickSpec {
                     expect(TestCancellable.cancelled).to( beTrue() )
                 }
             }
-        }
-    }
-}
-
-private extension String {
-    var URLEscapedString: String {
-        return self.stringByAddingPercentEncodingWithAllowedCharacters(NSCharacterSet.URLHostAllowedCharacterSet())!
-    }
-}
-
-private enum GitHub {
-    case Zen
-    case UserProfile(String)
-}
-
-extension GitHub : MoyaTarget {
-    var baseURL: NSURL { return NSURL(string: "https://api.github.com")! }
-    var path: String {
-        switch self {
-        case .Zen:
-            return "/zen"
-        case .UserProfile(let name):
-            return "/users/\(name.URLEscapedString)"
-        }
-    }
-    var method: Moya.Method {
-        return .GET
-    }
-    var parameters: [String: AnyObject]? {
-        return nil
-    }
-    var sampleData: NSData {
-        switch self {
-        case .Zen:
-            return "Half measures are as bad as nothing at all.".dataUsingEncoding(NSUTF8StringEncoding)!
-        case .UserProfile(let name):
-            return "{\"login\": \"\(name)\", \"id\": 100}".dataUsingEncoding(NSUTF8StringEncoding)!
-        }
-    }
-}
-
-private func url(route: MoyaTarget) -> String {
-    return route.baseURL.URLByAppendingPathComponent(route.path).absoluteString
-}
-
-private let failureEndpointClosure = { (target: GitHub) -> Endpoint<GitHub> in
-    let error = NSError(domain: "com.moya.error", code: 0, userInfo: [NSLocalizedDescriptionKey: "Houston, we have a problem"])
-    return Endpoint<GitHub>(URL: url(target), sampleResponseClosure: {.NetworkError(error)}, method: target.method, parameters: target.parameters)
-}
-
-private enum HTTPBin: MoyaTarget {
-    case BasicAuth
-
-    var baseURL: NSURL { return NSURL(string: "http://httpbin.org")! }
-    var path: String {
-        switch self {
-        case .BasicAuth:
-            return "/basic-auth/user/passwd"
-        }
-    }
-
-    var method: Moya.Method {
-        return .GET
-    }
-    var parameters: [String: AnyObject]? {
-        switch self {
-        default:
-            return [:]
-        }
-    }
-
-    var sampleData: NSData {
-        switch self {
-        case .BasicAuth:
-            return "{\"authenticated\": true, \"user\": \"user\"}".dataUsingEncoding(NSUTF8StringEncoding)!
         }
     }
 }
